requirements.txt
google-adk
google-generativeai
python-dotenv
yfinance